+++ app/routes/authorsignup.tsx
@@ -1,7 +1,5 @@
-import { Author, PrismaClient } from "@prisma/client"
+import { AuthorInput, createAuthor } from "~/service/Author"
 
-
-const prisma = new PrismaClient()
 
 /*
 
@@ -16,11 +14,7 @@ so I'll just create a branch includes some dummy data and tests to see if
 i can properly create the rows in prisma
 
 */
-type AuthorInput = {
-    name: string
-    biography: string
 
-}
 
 const dummyAuthor: AuthorInput = {
     name: "Kate DiCamillo",
@@ -29,17 +23,6 @@ const dummyAuthor: AuthorInput = {
 }
 
 
-// async
-export async function createAuthor(input: AuthorInput) {
-
-    const newUser = await prisma.author.create({
-        data: input
-    })
-
-    console.log("new user created:" + newUser)
-
-
-}
 
 export default function AuthorSignUp() {
     return (
